# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is prototype of aquafarm management service",
        "title": "Aquafarm Management Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/farms/{farmID}/ponds": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create new pond",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ponds"
                ],
                "summary": "Create new pond",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Farm ID",
                        "name": "farmID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pond Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PondResponse"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/ponds": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get all ponds with pagination and filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ponds"
                ],
                "summary": "Get all paginated ponds",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count of ponds each page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the sorts key of ponds sortation",
                        "name": "sorts",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the ponds search by code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "time",
                        "description": "the pond created date is filtered greather than or equal created date start",
                        "name": "createdDateStart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "time",
                        "description": "the pond created date is filtered less than or equal created date end",
                        "name": "createdDateEnd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PondPagingResponse"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create new pond",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ponds"
                ],
                "summary": "Create new pond",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Farm ID",
                        "name": "farmID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pond Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PondResponse"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreatePondRequest": {
            "type": "object",
            "required": [
                "code",
                "depth",
                "description",
                "loginData",
                "long",
                "wide"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "depth": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "farmCode": {
                    "type": "string"
                },
                "farmId": {
                    "type": "integer"
                },
                "loginData": {},
                "long": {
                    "type": "number"
                },
                "wide": {
                    "type": "number"
                }
            }
        },
...
        "dto.FarmResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "ponds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PondResponse"
                    }
                },
                "postalCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "village": {
                    "type": "string"
                }
            }
        },
        "dto.PondPagingResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "currentPage": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PondResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "sorts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "totalPage": {
                    "type": "integer"
                }
            }
        },
        "dto.PondResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "depth": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "farm": {
                    "$ref": "#/definitions/dto.FarmResponse"
                },
                "farmId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "long": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "wide": {
                    "type": "number"
                }
            }
        },
...
        "generateResp.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "generateResp.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}